+++ src/layers/DTGraphLayer.cpp
@@ -260,7 +260,10 @@ void DTGraphLayer::textChanged(CCTextInputNode* input){
     if (input == m_SessionSelectionInput->getInput() && m_DTLayer->m_SessionsAmount > 0){
         int selected = 1;
         if (input->getString() != "")
-            selected = std::stoi(input->getString());
+            try{
+                selected = std::stoi(input->getString());
+            }
+            catch (...) {}
         
         if (selected > m_DTLayer->m_SessionsAmount){
             selected = m_DTLayer->m_SessionsAmount;
@@ -283,7 +286,10 @@ void DTGraphLayer::textChanged(CCTextInputNode* input){
     if (input == m_RunSelectInput->getInput()){
         int selected = 0;
         if (input->getString() != "")
-            selected = std::stoi(input->getString());
+            try{
+                selected = std::stoi(input->getString());
+            }
+            catch (...) {}
 
         if (selected > 100){
             selected = 100;